move test_vocab_vectors_custom_cache (test.test_vocab.TestVocab)
